Use migration framework to trigger backfilling orphaned repositories (exist in db but not on disk)
